+++ e2e/internal/e2e/profile.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2019-2023, Sylabs Inc. All rights reserved.
+// Copyright (c) 2019-2025, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -258,7 +258,7 @@ func fakerootRequirements(t *testing.T) {
 	uid := uint32(origUID)
 
 	// check that current user has valid mappings in /etc/subuid
-	if _, err := fakeroot.GetIDRange(fakeroot.SubUIDFile, uid); err != nil {
+	if _, err := fakeroot.GetUIDRange(uid); err != nil {
 		t.Fatalf("fakeroot configuration error: %s", err)
 	}
 
@@ -267,7 +267,7 @@ func fakerootRequirements(t *testing.T) {
 	// *name*, it is keyed by user name, not by group name. This
 	// means that even if we are requesting the *group* mappings, we
 	// need to pass the *user* ID.
-	if _, err := fakeroot.GetIDRange(fakeroot.SubGIDFile, uid); err != nil {
+	if _, err := fakeroot.GetGIDRange(uid); err != nil {
 		t.Fatalf("fakeroot configuration error: %s", err)
 	}
 }
@@ -283,7 +283,7 @@ func ociRequirements(t *testing.T) {
 	uid := uint32(origUID)
 
 	// check that current user has valid mappings in /etc/subuid
-	if _, err := fakeroot.GetIDRange(fakeroot.SubUIDFile, uid); err != nil {
+	if _, err := fakeroot.GetUIDRange(uid); err != nil {
 		t.Fatalf("fakeroot configuration error: %s", err)
 	}
 
@@ -292,7 +292,7 @@ func ociRequirements(t *testing.T) {
 	// *name*, it is keyed by user name, not by group name. This
 	// means that even if we are requesting the *group* mappings, we
 	// need to pass the *user* ID.
-	if _, err := fakeroot.GetIDRange(fakeroot.SubGIDFile, uid); err != nil {
+	if _, err := fakeroot.GetGIDRange(uid); err != nil {
 		t.Fatalf("fakeroot configuration error: %s", err)
 	}
 }