+++ internal/pkg/fakeroot/config.go
@@ -17,7 +17,6 @@ import (
 	"strings"
 	"syscall"
 
-	specs "github.com/opencontainers/runtime-spec/specs-go"
 	"github.com/sylabs/singularity/v4/internal/pkg/util/user"
 	"github.com/sylabs/singularity/v4/pkg/util/fs/lock"
 )
@@ -292,11 +291,20 @@ func (c *Config) DisableUser(username string) error {
 	return nil
 }
 
+func (c *Config) getMappingEntries(user *user.User) []*Entry {
+	entries := make([]*Entry, 0)
+	for _, entry := range c.entries {
+		if entry.UID == user.UID {
+			entries = append(entries, entry)
+		}
+	}
+
+	return entries
+}
+
 // getUserEntry returns a user entry associated to a user and returns an error
-// if there is no entry for this user. If libsubid is true, entries are returned
-// from the subid file or via libsubid lookup. If libsubid is false, then
-// entries are only returned from the subid file.
-func (c *Config) getUserEntry(username string, libsubid bool) (*Entry, error) {
+// if there is no entry for this user.
+func (c *Config) GetUserEntry(username string) (*Entry, error) {
 	var largeRangeEntries []*Entry
 	entryCount := 0
 
@@ -305,10 +313,7 @@ func (c *Config) getUserEntry(username string, libsubid bool) (*Entry, error) {
 		return nil, fmt.Errorf("could not retrieve user information for %s: %s", username, err)
 	}
 
-	entries, err := c.getMappingEntries(u, libsubid)
-	if err != nil {
-		return nil, fmt.Errorf("failed to look up mapping entries for user %s: %w", username, err)
-	}
+	entries := c.getMappingEntries(u)
 
 	for _, entry := range entries {
 		if entry.invalid {
@@ -343,45 +348,4 @@ func (c *Config) getUserEntry(username string, libsubid bool) (*Entry, error) {
 		)
 	}
 	return nil, fmt.Errorf("no mapping entry found in %s for %s", c.file.Name(), username)
-}
-
-func (c *Config) GetUserEntry(username string) (*Entry, error) {
-	return c.getUserEntry(username, false)
-}
-
-func (c *Config) GetUserEntryWithLibSubid(username string) (*Entry, error) {
-	return c.getUserEntry(username, true)
-}
-
-// getPwUID is also used for mocking purpose
-var (
-	getPwUID = user.GetPwUID
-	getPwNam = user.GetPwNam
-)
-
-// GetIDRange determines UID/GID mappings based on configuration
-// file provided in path, and libsubid if compiled with support.
-func GetIDRange(path string, uid uint32) (*specs.LinuxIDMapping, error) {
-	config, err := GetConfig(path, false, getPwNam)
-	if err != nil {
-		return nil, err
-	}
-	defer config.Close()
-
-	userinfo, err := getPwUID(uid)
-	if err != nil {
-		return nil, fmt.Errorf("could not retrieve user with UID %d: %s", uid, err)
-	}
-	e, err := config.GetUserEntryWithLibSubid(userinfo.Name)
-	if err != nil {
-		return nil, err
-	}
-	if e.disabled {
-		return nil, fmt.Errorf("your fakeroot mapping has been disabled by the administrator")
-	}
-	return &specs.LinuxIDMapping{
-		ContainerID: 1,
-		HostID:      e.Start,
-		Size:        e.Count,
-	}, nil
 }